order_service/requirements.txt
grpcio==1.59.3
grpcio-tools==1.59.3
motor==3.3.2
pymongo==4.6.1
protobuf==4.25.1
termcolor==2.3.0
